requirements.txt
streamlit
plotly
pandas
requests
python-dateutil
cryptography
python-dotenv
websockets
